[marv] Properly set and reset outdated state of datasets
So far, datasets were marked outdated, if the mtime of any file
changed, independent of whether there are actually node runs, and so
far the state got never reset.

Now, a dataset is outdated if any of its current store output is older
than the newest of its files, and the outdated state is evaluated if
the mtime of a file changes and after details are rendered for a
dataset, which always happens if a node was run.