Fix formatter issue in case of pushing no value in a node
Assume that a MARV node never pushes a value. This e.g. is the case if the node has a topic as an input but only some bags contain this topic. For the other bags, MARV will produce a Python error.

This error can be easily fixed by taking into account that a node can push no value. If no value is pushed, `calltree` will return `None`. Thus we have to account for `None` in all formatters.