[signoff] Handle committer being reset after rebase
GitLab offers rebasing branches through the web. After this, as when
doing a rebase manually, the committer is set to the person performing
the rebase and the original committer is lost.

See docstring of cli function for the now implemented checks. This
does not replace a manual inspection of every commit.

This also adds a selftest subcommand.