Replace testing make_sink with run_nodes wrapper
The main purpose of make_sink was to place a list of values on the node
instance to be compared to expected values by tests. This is now
achieved by gathering values for the selected nodes into a dictionary,
which is provided when run_nodes is run as part of tests.